Moved all the shiny new QP controller functions into actual libraries
