+++ systems/controllers/encodeQPInputLCMMex.cpp
@@ -1,6 +1,6 @@
 #include <lcm/lcm-cpp.hpp>
 #include "drake/lcmt_qp_controller_input.hpp"
-#include "controlUtil.h"
+#include "QPCommon.h"
 
 using namespace std;
